+++ spark-wx-shop/spark-wx-shop-biz/src/main/java/com/spark/platform/wx/shop/biz/specs/service/ShopSpecsAttrService.java
@@ -17,5 +17,5 @@ public interface ShopSpecsAttrService extends IService<ShopSpecsAttr> {
      * @param id
      * @return
      */
-    ShopSpecsAttr getShopSpecsAttr(Long id);
+    ShopSpecsAttr getShopSpecsAttr(Integer id);
 }
